meta-qti-bsp: Set the dependency for recovery.service
In case of apq8017-qsap, recovery service is not able
to read the update package stored on /data as /data
partition is notmounted by the time recovery.service
is started.
By adding the dependency on mount-data.service, we
are making sure that /data partition is mounted
before recovery.service gets started.

Change-Id: I1b2023694672a806e2e496a8b5a9ecba9f7d3f29